weight fixes and ui enhancements

+++ src/Vehicle.cpp
@@ -202,7 +202,7 @@ MapIcon *godot::Vehicle::get_move_waypoint() {
 	return this->moveWaypoint;
 }
 
-uint16_t godot::Vehicle::get_weight() {
+float godot::Vehicle::get_weight() {
 	return this->weight;
 }
 
@@ -264,7 +264,7 @@ void godot::Vehicle::set_move_waypoint(MapIcon *waypoint) {
 	this->moveWaypoint = waypoint;
 }
 
-void godot::Vehicle::set_weight(uint16_t value) {
+void godot::Vehicle::set_weight(float value) {
     this->weight = value;
 }
